lf polish-gate: cli: refactor flags and move cli code to dedicated modules
- Rename -v/--paste to -c/--clipboard, -L/--voice to -v/--voice
- Move lf and lfd CLI code from __init__.py to cli.py
- Support flags without step name (lf -m codex, lf --web)
- Update docs to match current command syntax
- Document lfops worktree and abandon commands
