+++ README.md
@@ -52,7 +52,7 @@ lfops pr
 Define goals, review PRs when you wake.
 
 ```bash
-lfd loop src/ --flow ship
+lfd loop ship src/
 ```
 
 ![loops demo](docs/loops-demo.gif)
